requirements.txt
# --- LangGraph Core ---
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langgraph-cli[inmem]

# --- LangChain ---
langchain-core
langchain-openai
langchain-ollama
langchain-anthropic
langchain-community

# --- LLM y modelos ---
ollama
llama-index==0.11.0
litellm

# --- Phoenix y OpenInference ---
arize-phoenix
arize-phoenix-otel
openinference-instrumentation
openinference-instrumentation-langchain
openinference-instrumentation-openai
openinference-instrumentation-llama_index
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-fastapi

# --- Evaluación y visualización ---
ipywidgets
matplotlib
seaborn
pandas
tqdm
codecarbon
nvidia-ml-py3

# --- SQL y almacenamiento ---
sqlalchemy
aiosqlite
duckdb

# --- Utilidades ---
notebook
python-dotenv
httpx
numpy
trustcall
openai
openai-agents